requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-filter==23.3
python-dotenv==1.0.0